requirements.txt
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.3
cachetools==4.2.2
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
defusedxml==0.7.0rc1
Django==3.1
django-autocomplete-light==3.8.1
django-cleanup==5.0.0
django-filter==2.4.0
django-grappelli==2.14.2
django-guardian==2.3.0
django-modeltranslation==0.15.1
django-templated-mail==1.1.1
django-tinymce==3.0.2
django-user-agents==0.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
google-api-core==1.31.0
google-api-python-client==2.12.0
google-auth==1.32.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
grappelli-modeltranslation==0.1.3
httplib2==0.19.1
idna==2.10
isort==4.3.21
itypes==1.2.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
Markdown==3.2.2
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.1.0
packaging==21.0
Pillow==7.2.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.7.2
PyJWT==1.7.1
pylint==2.5.3
pyparsing==2.4.7
python3-openid==3.2.0
pytz==2020.1
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sorl-thumbnail==12.6.3
sqlparse==0.3.1
toml==0.10.1
ua-parser==0.10.0
uritemplate==3.0.1
urllib3==1.26.2
user-agents==2.2.0
wrapt==1.12.1
